Simplification of buildings
Buildings are simplified based on a score system. The score respresents
the maximum squared distance of two successive points of polygon of a
building.
In the parent node the buildings with the lowest score will be
simplified till the max amount of triangles has been reached. A building
is simplified by combining the points associated with the score